+++ crates/wasm-bridge-cli/src/transformers/mod.rs
@@ -1,9 +1,0 @@
-mod js_transform;
-mod transformer;
-mod universal;
-mod version;
-
-pub use js_transform::*;
-pub use transformer::*;
-pub use universal::*;
-pub use version::*;