+++ crates/wasm-bridge-js/src/component/component.rs
@@ -3,12 +3,8 @@ use std::{borrow::Borrow, collections::HashMap, rc::Rc};
 use anyhow::Context;
 use js_sys::{Function, Uint8Array, WebAssembly};
 use wasm_bindgen::prelude::*;
-use zip::ZipArchive;
 
-use crate::{
-    helpers::{self, map_js_error},
-    AsContextMut, DropHandler, Engine, Result,
-};
+use crate::{helpers::map_js_error, AsContextMut, DropHandler, Engine, Result};
 
 use super::*;
 
@@ -21,37 +17,8 @@ pub struct Component {
 
 impl Component {
     pub fn new(_engine: &Engine, bytes: impl AsRef<[u8]>) -> Result<Self> {
-        Self::from_zip(&bytes).or_else(|_| {
-            let loader = ComponentLoader::new().unwrap(); // TODO: bad unwrap
-            loader.compile_component(bytes.as_ref())
-        })
-    }
-
-    pub fn from_zip(zip_bytes: impl AsRef<[u8]>) -> Result<Self> {
-        let cursor = std::io::Cursor::new(zip_bytes);
-        let mut archive = ZipArchive::new(cursor)?;
-
-        let mut version = Option::<String>::None;
-
-        let mut files = vec![];
-
-        for i in 0..archive.len() {
-            let mut file = archive.by_index(i)?;
-            let filename = file.name().to_owned();
-
-            let mut file_bytes = Vec::<u8>::with_capacity(file.size() as usize);
-            std::io::copy(&mut file, &mut file_bytes)?;
-
-            if filename.ends_with("version.txt") {
-                version = Some(String::from_utf8(file_bytes)?);
-            } else {
-                files.push((filename, file_bytes));
-            }
-        }
-
-        Self::check_version(version.as_deref());
-
-        Self::from_files(files)
+        let loader = ComponentLoader::new().context("create new component loader")?;
+        loader.compile_component(bytes.as_ref())
     }
 
     pub(crate) fn from_files(files: Vec<(String, Vec<u8>)>) -> Result<Self> {
@@ -60,13 +27,13 @@ impl Component {
 
         for (filename, file_bytes) in files.into_iter() {
             if filename.ends_with(".wasm") {
-                wasm_cores.insert(filename, file_bytes); // TODO: remove folder from filename?
+                wasm_cores.insert(filename, file_bytes);
             } else if filename.ends_with("sync_component.js") {
                 instantiate = Some(Self::load_instantiate(&file_bytes)?);
             }
         }
 
-        let instantiate = instantiate.context("component js file not found in zip")?;
+        let instantiate = instantiate.context("component js file not found in files")?;
 
         let (compile_core, drop0) = Self::make_compile_core(wasm_cores);
         let (instantiate_core, drop1) = Self::make_instantiate_core();
@@ -112,27 +79,11 @@ impl Component {
         Ok(instantiate)
     }
 
-    fn check_version(version: Option<&str>) {
-        match version {
-            Some(version) => {
-                let current_version = env!("CARGO_PKG_VERSION");
-                if version != current_version {
-                    helpers::warn(&format!(
-                        "Version mismatch: {} {version}, {} {current_version}",
-                        "component was build with wasm-bridge-cli version",
-                        "but wasm-bridge is running version"
-                    ));
-                }
-            }
-            None => helpers::warn("Missing version file in component zip."),
-        }
-    }
-
     fn make_compile_core(wasm_cores: HashMap<String, Vec<u8>>) -> (JsValue, DropHandler) {
         let closure = Closure::<dyn Fn(String) -> WebAssembly::Module>::new(move |name: String| {
             let bytes = wasm_cores.get(&name).unwrap();
             let byte_array = Uint8Array::from(bytes.borrow());
-            WebAssembly::Module::new(&byte_array.into()).unwrap()
+            WebAssembly::Module::new(&byte_array.into()).unwrap() // TODO: user error
         });
 
         DropHandler::from_closure(closure)
@@ -148,9 +99,4 @@ impl Component {
 
         DropHandler::from_closure(closure)
     }
-}
-
-pub fn new_universal_component(engine: &Engine, bytes: impl AsRef<[u8]>) -> Result<Component> {
-    // Component::from_zip(bytes)
-    Component::new(engine, bytes)
 }